Implementação de metodos nos controllers e exceptions

+++ Compayz/src/main/java/com/compayz/domain/produto/InfoProduto.java
@@ -10,5 +10,5 @@ public class InfoProduto{
 
 	private String descricao;
 	private BigDecimal valor;
-	private Long quantidade;
+	private Integer quantidade;
 }
